+++ boards/feather_m0/examples/ssd1306_graphicsmode_128x64_spi.rs
@@ -69,7 +69,6 @@ use bsp::{entry, periph_alias, pin_alias};
 use hal::clock::GenericClockController;
 use hal::delay::Delay;
 use hal::prelude::*;
-use hal::time::MegaHertz;
 use pac::{CorePeripherals, Peripherals};
 
 #[entry]
@@ -89,7 +88,7 @@ fn main() -> ! {
     let spi_sercom = periph_alias!(peripherals.spi_sercom);
     let spi = bsp::spi_master(
         &mut clocks,
-        MegaHertz(10),
+        10.MHz(),
         spi_sercom,
         &mut peripherals.PM,
         pins.sclk,